test(admin-form.controller.spec): shifts over associated tests
